# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Business Suite Automation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get all inventory items
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [Summer, Winter, Spring, Girls, Boys]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create a new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a single inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
...
components:
  schemas:
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        productId:
          type: string
        category:
          type: string
          enum: [Summer, Winter, Spring]
        gender:
          type: string
          enum: [Girls, Boys, both]
        ageMin:
          type: integer
        ageMax:
          type: integer
        price:
          type: number
        stock:
          type: integer
        descriptionAr:
          type: string
          nullable: true
        descriptionEn:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        parentProductId:
          type: string
          nullable: true
        colorVariantName:
          type: string
          nullable: true
        colors:
          type: string
          nullable: true
        available:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - nameAr
        - nameEn
        - productId
        - category
        - ageMin
        - ageMax
        - price
        - stock
        - available
        - createdAt
        - updatedAt
    CreateInventoryItemBody:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        productId:
          type: string
        category:
          type: string
          enum: [Summer, Winter, Spring]
        gender:
          type: string
          enum: [Girls, Boys, both]
        ageMin:
          type: integer
        ageMax:
          type: integer
        price:
          type: number
        stock:
          type: integer
        descriptionAr:
          type: string
          nullable: true
        descriptionEn:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        parentProductId:
          type: string
          nullable: true
        colorVariantName:
          type: string
          nullable: true
        colors:
          type: string
          nullable: true
        available:
          type: boolean
      required:
        - nameAr
        - nameEn
        - productId
        - category
        - ageMin
        - ageMax
        - price
        - stock
    UpdateInventoryItemBody:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        productId:
          type: string
        category:
          type: string
          enum: [Summer, Winter, Spring]
        gender:
          type: string
          enum: [Girls, Boys, both]
        ageMin:
          type: integer
        ageMax:
          type: integer
        price:
          type: number
        stock:
          type: integer
        descriptionAr:
          type: string
          nullable: true
        descriptionEn:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        parentProductId:
          type: string
          nullable: true
        colorVariantName:
          type: string
          nullable: true
        colors:
          type: string
          nullable: true
        available:
          type: boolean
...